+++ ext/opentelemetry-ext-psycopg2/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+## 0.8b0
+
+Released 2020-05-27
+
 - Implement instrumentor interface, enabling auto-instrumentation ([#694]https://github.com/open-telemetry/opentelemetry-python/pull/694)
 
 ## 0.4a0